requirements.txt
streamlit
openai
faiss-cpu
numpy
tiktoken
python-dotenv
pyarrow
pandas
lxml
tqdm
requests 